requirements.txt
openai
nemoguardrails
langchain
pyyaml
python-dotenv
